fix(sc2/emulated): calibrate trainTimeInLoops to integer loops (#149)
SC2Data.trainTimeInLoops(UnitType) changed from double to int. Float
literals (e.g. 268.8 for Probe) caused floor division to produce 12
ticks instead of 13 for loop offsets [14, 17], giving firstUnitDivergenceTick
= 86. Empirical values (PROBE=272, ZEALOT=618, STALKER=698) calibrated from
29 AI Arena replays via SC2TrainTimeCalibrationTest (range-bounded modal
algorithm, two-source cross-validation). EmulatedGame.startTraining formula
is now pure integer arithmetic. ReplayValidationTest tightened: first
divergence now at tick 150, sole remaining cause is vespene gap (#148).

Closes #149

docs(issue-149-completesat-rounding): apply design journal

+++ src/main/java/io/quarkmind/sc2/emulated/EmulatedGame.java
@@ -300,7 +300,7 @@ public class EmulatedGame {
         boolean isEnemy  = (state == enemy);
         int  loopOffset  = (int)(absLoop % SC2Data.LOOPS_PER_TICK);
         long completesAt = gameFrame
-            + (int)((loopOffset + SC2Data.trainTimeInLoops(unitType)) / SC2Data.LOOPS_PER_TICK);
+            + (loopOffset + SC2Data.trainTimeInLoops(unitType)) / SC2Data.LOOPS_PER_TICK;
         state.buildingTrainingUntil.put(buildingTag, completesAt);
         state.pendingCompletions.add(new PlayerState.PendingCompletion(completesAt, () -> {
             state.buildingTrainingUntil.remove(buildingTag);
